refactor: refactored paymentprocessor
