+++ services/paymentprocessor/core/clients.py
@@ -1,0 +1,3 @@
+import httpx
+
+http_client: httpx.AsyncClient = None
